Move AddressHelper to a concern
