+++ spec/support/shared_examples/forms/address_lookup_form.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 RSpec.shared_examples "an address lookup form", vcr: true do |form_factory|
-  it "includes AddressHelper" do
+  it "includes CanAddOrReplaceAnAddress" do
     included_modules = described_class.ancestors.select { |ancestor| ancestor.instance_of?(Module) }
-    expect(included_modules).to include(WasteExemptionsEngine::AddressHelper)
+    expect(included_modules).to include(WasteExemptionsEngine::CanAddOrReplaceAnAddress)
   end
 
   it "is a type of AddressLookupForm" do
